refactor rename justMarketplace.ts -> marketplace.ts
