+++ src/contentScript/sidebarActivation.ts
@@ -32,7 +32,7 @@ import {
 } from "@/background/messenger/external/_implementation";
 import reportError from "@/telemetry/reportError";
 import { reportEvent } from "@/telemetry/events";
-import { getInstalledRecipeIds } from "@/contentScript/justMarketplace";
+import { getInstalledRecipeIds } from "@/contentScript/marketplace";
 
 async function isUserLoggedIn(): Promise<boolean> {
   const authHeaders = await getAuthHeaders();
